make tests compile
